### 2021-05-09 #3266 @arfon
@jbytecode - thanks for your submission to JOSS. We're currently working through a backlog of submissions so I'm marking yours as _waitlisted_ for now. We'll assign an editor to handle your paper as and when a suitable editor becomes available. Thanks for your patience!
