### 2021-05-09 388cb2c
update electre bibtex entry in paper
