+++ paper/paper.bib
@@ -113,7 +113,8 @@ and Vansnick, Jean-Claude},
 }
 
 
-@misc{electre,
+@article{electre,
+  doi       = {10.1051/ro/196802v100571},
   author    = {Roy, B.},
   journal   = {RAIRO - Operations Research - Recherche Opérationnelle},
   language  = {fre},
