### 2021-05-31 #3266 @jbytecode
@kyleniemeyer The invited editor's notifications may be turned off.
