### 2021-06-15 #3266 @drvinceknight
> @kyleniemeyer The invited editor's notifications may be turned off.

Not quite turned off but overflowing 😆 Apologies for my delay on this, I'll happily take it on.
