### 2021-06-15 #3266 @jbytecode
Dear editor @drvinceknight

I am listing some potential reviewers (from bottom of the list) with their research areas. Julia is in their preferred or other language list.

bonh: numerical mathematics
AvianaGlobal: statistics and economics
degleris1: optimization and statistics
brunaw: statistics
sylvaticus: social sciences
salleuska: statistics and optimization
mvaldes: optimization and numerical methods
ayushpatnaikgit: statistics
jbcaillau: optimization
justinsgray: numerical optimization
