### 2021-06-18 #3266 @drvinceknight
> Hi @drvinceknight! I would very much like to review this submission. However, I feel obligated to bring to your attention, that in April I reviewed another JOSS submission during which I asked for an editorial scope check. Afterwards the editor and I decided to remove myself from the review process due to insufficient knowledge in the specific field of that submission. Unfortunately, the submission is still flagged for editorial review - so it is not done yet and I feel uncomfortable to start another review while that one still lingers around. My point is, that I might not have the heart to ask for another scope check (or some other editorial stuff) during the review of this submission even if I feel it might be needed - after all we all do this on the side and I greatly appreciate the work of the editors! I hope you get what I mean.

Thank you @bonh for responding. No problem if you cannot review this submission.

If you can let me know what the other review is I'll chase up to try and make sure it's not a problem for you.
