### 2021-06-29 #3266 @drvinceknight
@sylvaticus @salleuska @mvaldes @ayushpatnaikgit I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "JMcDM: A Julia package for multiple-criteria decision-making tools". You can find more information at the top of this Github issue.
