### 2021-06-29 #3266 @sylvaticus
Hello, I could do it, but I am confident only with a small subset of the methods implemented in the package.
