### 2021-06-29 #3430 @whedon
[REVIEW]: JMcDM: A Julia package for multiple-criteria decision-making tools
**Submitting author:** @jbytecode (<a href="http://orcid.org/0000-0002-9402-1982">Mehmet Hakan Satman</a>)
**Repository:** <a href="https://github.com/jbytecode/JMcDM" target ="_blank">https://github.com/jbytecode/JMcDM</a>
**Version:** v0.2.4
