### 2021-07-01 #3430 @sylvaticus
**Sylvaticus review**

The software follows the Julia standards for registered packages: it is easy to install, it has a clear licence, the methods are documented, and an extensive testing system covers much of the software.
I have a few minor/specific points concerning the documentation, the API and the JOSS paper in the sections that follows.

I have also two more general comments.

The first one is that while in the paper several other "similar" software are cited, still it is not clear the innovation bring by this package. Is it easier to use than the cited software? Is it more complete in terms of methods ? Is it faster for large problems (here a benchmark would be ideal) ? I believe that the package would benefit from a little more detailed "statement of needs" (in the paper and/or in the online documentation)

The second general comment is somehow related. While it is impressive the number of methods supported, it is a bit like a list of completely unrelated tools.
I believe that in both the documentation and in the API design would be better to look for a taxonomy/way to group/organise the different methods.
One possible suggestion could be to have an interface function like `MCDM(df,w,fnd;method=Topsis())` that wraps the various methods. In this way, the user would have to remember only one function and then eventually with `?MCDM` get more info about the supported methods, rather than browse the documentation for the supported functions.
The function would have an optional `method` keyword that receive a method-specific structure with its parameters (e.g. `v` for `vikor`, or `zeta` for `grey`).
Also, the triplet `(df,w,fnd)` could be documented more in detail, but only once, and then each specific method would have its own documentation.
The same for the output… most methods return a structure with `decisionMatrix`, `weights`, `scores`, `ranking`, and `bestIndex`. I believe it should be better to create a common output with these fields (where they could be documented once) plus an object carrying the variable information. Note that using the suggested approach the `MCDM` function would remain type stable, as each method would be a different type, so the compiler would still specialise each "method".
Document the `summary` function could also be useful as well (if exist) some indications on when one method should be preferable to the others.

Following are my specific points.

### Minor/specific points concerning the software

- the optional parameters, like `v` in vikor and `zeta` in `grey` are sometimes positional argument sometimes keyword argument
- in `promethee` maybe it is better to keep the first 3 parameters as in the other methods dm, w, fns and then add the specific argument parameter
- use a default (all methods? the "recommended" ones ?) for the `summary` function
- implement a `print` method for the results
- What the function `makeminmax` does ? It seems it is not needed:

```
julia> fns2 = [minimum, maximum, minimum, minimum, maximum]
5-element Vector{Function}:
 minimum (generic function with 18 methods)
 maximum (generic function with 17 methods)
 minimum (generic function with 18 methods)
 minimum (generic function with 18 methods)
 maximum (generic function with 17 methods)

julia> fns = makeminmax([minimum, maximum, minimum, minimum, maximum])
5-element Vector{Function}:
 minimum (generic function with 18 methods)
 maximum (generic function with 17 methods)
 minimum (generic function with 18 methods)
 minimum (generic function with 18 methods)
 maximum (generic function with 17 methods)

julia> fns == fns2
true
```
- What the undocumented `makeDecisionMatrix` function does other than transforming a matrix to a df ?
- When apply the data of the paper to the `moora` function, I have all decisions with the same score. Maybe there should be a warning or a flag when the algorithm doesn't converge, or `bestIndex` should be rather be `bestIndices` with a singleton being a particular case (not only here in the `moora` function, but more in general)

### Minor/specific points concerning the text of the JOSS paper

- L1 (title): I don't see very well the "for" word going well with the "tools" word... maybe "A Julia package for multiple-criteria decision-making" or "A Julia collection of multiple-criteria decision-making tools"...
- L11-13: I am not sure "the package" refers to Julia or to this package in particular. In the first case maybe better "The language", in the second case it is more to "use" rather than to "write" recently published methods. Also, you use the words "multiple-criteria decision-making" but then you use the acronym for "multiple-criteria decision-analysis" (MCDA).
- L 10,52,71,78: "REPL" is cited (without definition) 4 times. I don't think it is necessary to focus so much on the REPL
- L 55-64: Maybe order these methods, alphabetically or by year ??
- I believe there is a misunderstanding with the "State of the field" section. While I like its content and I think it is important to give an idea of the subject covered in the software, I think that the section with that title is more for comparing the proposed package with the existing alternatives (from the reviewer checklist: "State of the field: Do the authors describe how this software compares to other commonly-used packages?")

### Minor points concerning the online documentation
- It is useful to have the documentation for the stable (and all the different versions) and the development version separated. This can be done automatically using Documenter, see here for [a tutorial](https://syl1.gitbook.io/julia-language-a-concise-tutorial/language-core/11-developing-julia-packages#documenting-our-package) and associated ["dummy" package](https://github.com/sylvaticus/MyAwesomePackage.jl/)
- I find also useful to have a link from the documentation back to the main page of the github repo (the "edit to Github" sends to a inner page")
- Also, while the testing seems extensive, having a coverage section would help to highlight the parts that are still not covered by the tests. Again, refer to the same page for a template to add automatic coverage reports.
- JMcDM.electre function documentation: the output is copy/pasted from the `topsis` function above, but it has not been adapted. What are `C` and `D` ?
- `damatel` func documentation: _comparisonMat::Array{Float,2}: n × m matrix of input values._ What are the input values ?  Also doc say Matrix, but the function wants a df. I can't get this function working with the df in the paper. It seems the matrix must be squared, or a dimension mismatch error happens.
- `AHP` documentation: the arguments of the function in the given example is (df, vector of matrix), while in the docstring both are df
- `promethee` docstring: the conversion of prefs is not needed
- Some useful functions, like `summary`, are undocumented
- The "community guidelines" for contributing to the software are missing
