### 2021-07-01 180ede8
add mcdm(df, w, fns, method) type standard dispatch for mcdm methods
