### 2021-07-01 78dfcd8
add print(T1 <: MCDMResult) method for all MCDM methods
