+++ src/.dev/playground.jl
@@ -19,6 +19,7 @@ using Statistics
 
 include("../types.jl")
 include("../mcdm.jl")
+include("../print.jl")
 include("../utilities.jl")
 include("../topsis.jl")
 include("../vikor.jl")
