+++ src/print.jl
@@ -1,0 +1,157 @@
+function print(result::TopsisResult)
+    println("Scores:")
+    println(result.scores)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::ARASResult)
+    println("Scores:")
+    println(result.scores)
+    println("Orderings: ")
+    println(result.orderings)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::CODASResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+
+function print(result::COPRASResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::CRITICResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::CoCoSoResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::EDASResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::ElectreResult)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::GreyResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ordering: ")
+    println(result.ordering)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::MABACResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::MAIRCAResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::MARCOSResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::MooraResult)
+    println("Scores:")
+    println(result.scores)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::PrometheeResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::SawResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::VikorResult)
+    println("Scores:")
+    println(result.scores)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::WASPASResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+function print(result::WPMResult)
+    println("Scores:")
+    println(result.scores)
+    println("Ranking: ")
+    println(result.ranking)
+    println("Best indices:")
+    println(result.bestIndex)
+end
+
+
+
+
+
