### 2021-07-01 6fe8f58
add method(setting::MCDMSetting) type dispatch
