### 2021-07-01 dec1f57
add doc for MCDMSetting, fix promethee test
