+++ src/summary.jl
@@ -1,3 +1,48 @@
+"""
+        summary(decisionMat, weights, fns, methods)
+
+Apply more methods for a given decision problem. The methods accept standart number of arguments.   
+
+# Arguments:
+ - `decisionMat::DataFrame`: n × m matrix of objective values for n candidate (or strategy) and m criteria 
+ - `weights::Array{Float64, 1}`: m-vector of weights that sum up to 1.0. If the sum of weights is not 1.0, it is automatically normalized.
+ - `fns::Array{Function, 1}`: m-vector of function that are either minimize or maximize.
+ - `methods::Array{Symbol, 1}`: Array of symbols. The elements can be :topsis, :electre, :cocoso, :copras, :moora, :vikor, :grey, :aras, :saw, :wpm, :waspas, :edas, :marcos, :mabac, :mairca, :copras, :critic
+
+# Description 
+    This method outputs a summarized output using more than MCDM methods in a comparable way. 
+
+    # Output 
+- `::DataFrame`: A DataFrame object, methods in columns, and alternatives in rows. Green check symbol indicates the selected alternative as the best by the corresponding method.
+
+# Examples
+```julia-repl
+julia> df = DataFrame(
+:age        => [6.0, 4, 12],
+:size       => [140.0, 90, 140],
+:price      => [150000.0, 100000, 75000],
+:distance   => [950.0, 1500, 550],
+:population => [1500.0, 2000, 1100]);
+
+
+julia> methods = [:topsis, :electre, :vikor, :moora, :cocoso, :wpm, :waspas]
+
+julia> w  = [0.036, 0.192, 0.326, 0.326, 0.12];
+
+julia> fns = [maximum, minimum, maximum, maximum, maximum];
+
+
+julia> result = summary(df, w, fns, methods)
+3×7 DataFrame
+ Row │ topsis  electre  cocoso  moora   vikor   wpm     waspas 
+     │ String  String   String  String  String  String  String 
+─────┼─────────────────────────────────────────────────────────
+   1 │                           ✅      ✅
+   2 │  ✅      ✅       ✅                      ✅      ✅
+   3 │
+
+```
+"""
 function summary(
     decisionMat::DataFrame, 
     weights::Array{Float64,1}, 
@@ -92,11 +137,6 @@ function summary(
         resultdf[:,:copras] = map(x -> if result.bestIndex == x check else " " end, 1:nalternatives)
     end 
 
-    if :cocoso in methods
-        result = cocoso(decisionMat, weights, fns)
-        resultdf[:,:cocoso] = map(x -> if result.bestIndex == x check else " " end, 1:nalternatives)
-    end 
-
     if :critic in methods
         result = critic(decisionMat, fns)
         resultdf[:,:critic] = map(x -> if result.bestIndex == x check else " " end, 1:nalternatives)
@@ -106,3 +146,15 @@ function summary(
     return resultdf
 
 end
+
+
+
+
+function summary(setting::MCDMSetting, methods::Array{Symbol,1})
+    summary(
+        setting.df,
+        setting.weights,
+        setting.fns,
+        methods
+    )
+end 
