+++ docs/src/index.md
@@ -1,4 +1,6 @@
 # Contents 
+
+[The GitHub Repository](https://github.com/jbytecode/JMcDM)
 ```@contents
 Pages = ["index.md", "mcdms.md", "game.md"]
 Depth = 3
