### 2021-07-01 330ce71
Delete codecov.yml
