+++ .github/workflows/main.yml
@@ -35,6 +35,3 @@ jobs:
             ${{ runner.os }}-
       - uses: julia-actions/julia-buildpkg@latest
       - uses: julia-actions/julia-runtest@latest
-
-    after_success:
-    - julia -e 'using Pkg; cd(Pkg.dir("JMcDM")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
