### 2021-07-01 c3dabb4
Delete codecov.yml
