+++ .github/workflows/codecov.yml
@@ -1,2 +1,0 @@
-codecov:
-  token: 19fddd76-3ad1-41f6-8138-5c88df5b98db
