### 2021-07-01 becba21
Create .travis.yml
