+++ .travis.yml
@@ -2,4 +2,4 @@ language: julia
 julia:
   - 1.6
  
- codecov: true
+codecov: true
