+++ .travis.yml
@@ -1,17 +1,0 @@
-language: julia
-julia:
-  - 1.6
- 
-codecov: true
-
-script:
-  - julia --project --check-bounds=yes -e 'import Pkg; Pkg.build(); Pkg.test("JMcDM"; coverage=true)'
-
-# After successful build submit coverage report and deploy updated documentation
-# But limit deployment to the linux worker so this is done only once.
-after_success:
-  # Install required packages for coverage and documentation
-  - julia --project -e 'import Pkg; Pkg.add("Coverage");'
-
-  # Submit test coverage report
-  - julia --project -e 'using Coverage; Coveralls.submit(Coveralls.process_folder())'
