### 2021-07-01 715e692
documentations for ahp and electre fixed
