### 2021-07-02 6f59b15
add tests for topsis and vikor
