### 2021-07-02 7acee7d
add tests for implemented methods
