+++ src/.dev/test.jl
@@ -123,6 +123,11 @@ end
 	@test isa(result2, TopsisResult)
 	@test result2.bestIndex == result.bestIndex
 	@test result.scores == result2.scores
+
+    result3 = mcdm(setting, TopsisMethod())
+    @test isa(result3, TopsisResult)
+	@test result3.bestIndex == result.bestIndex
+	@test result3.scores == result.scores
 end
 
 @testset "VIKOR" begin
@@ -151,6 +156,10 @@ end
 	result2 = vikor(setting)
 	@test result.scores == result2.scores
 	@test result.bestIndex == result2.bestIndex
+
+    result3 = mcdm(setting, VikorMethod())
+    @test result.scores == result3.scores
+    @test result.bestIndex == result3.bestIndex 
 end
 
 
@@ -173,6 +182,15 @@ end
     
     @test isapprox(result.C, [0.36936937,  0.01501502, -2.47347347,  2.08908909], atol=tol)
     @test isapprox(result.D, [0.1914244, -0.1903929,  2.8843076, -2.8853391], atol=tol)
+
+    setting = MCDMSetting(dmat, w, fns)
+    result2 = electre(setting)
+    @test result isa ElectreResult
+    @test result.bestIndex == result2.bestIndex
+    
+    result3 = mcdm(setting, ElectreMethod())
+    @test result3 isa  ElectreResult
+    @test result3.bestIndex == result.bestIndex
 end
 
 
@@ -194,6 +212,17 @@ end
     @test result.bestIndex == 4
 
     @test isapprox(result.scores, [0.33159387, 0.29014464, 0.37304311, 0.01926526], atol=tol)
+
+    setting = MCDMSetting(dmat, w, fns)
+    result2 = moora(setting)
+    @test result2 isa MooraResult
+    @test result2.bestIndex == result.bestIndex
+    @test result2.scores == result.scores
+    
+    result3 = mcdm(setting, MooraMethod())
+    @test result3 isa MooraResult
+    @test result3.bestIndex == result.bestIndex
+    @test result3.scores == result.scores
 end
 
 
@@ -711,6 +740,17 @@ end
     0.5464285714285715, 0.5762820512820512, 0.650952380952381], atol=tol)
 
     @test result.bestIndex == 2
+
+    setting = MCDMSetting(df, w, functionlist)
+    result2 = grey(setting)
+    @test result2 isa GreyResult
+    @test result2.scores == result.scores 
+    @test result2.bestIndex == result.bestIndex
+
+    result3 = mcdm(setting, GreyMethod())
+    @test result3 isa GreyResult
+    @test result3.scores == result.scores 
+    @test result3.bestIndex == result.bestIndex
 end
 
 
@@ -750,6 +790,17 @@ end
         @test isapprox(result.scores, [0.553228, 0.713485, 0.837428, 0.514657, 0.579342], atol=tol)
         @test result.bestIndex == 3
         @test result.ranking == [3, 2, 5, 1, 4]
+
+        setting = MCDMSetting(df, weights, fns)
+        result2 = saw(setting)
+        @test result2 isa SawResult
+        @test result2.scores == result.scores 
+        @test result2.bestIndex == result.bestIndex
+
+        result3 = mcdm(setting, SawMethod())
+        @test result3 isa SawResult
+        @test result3.scores == result.scores 
+        @test result3.bestIndex == result.bestIndex
     end
 end
 
@@ -771,13 +822,22 @@ end
     w = [0.05, 0.20, 0.10, 0.15, 0.10, 0.40]
 
     result = aras(df, w, functionlist)
-
     @test isa(result, ARASResult)
-    
[…]

[…]
