+++ src/grey.jl
@@ -1,10 +1,10 @@
 """
-        grey(decisionMat, weights, fs, zeta)
+        grey(decisionMat, weights, fs; zeta)
 
     Perform GRA (Grey Relational Analysis) for a given decision matrix and weights.
 
 # Arguments:
- - `decisionMat::Array{Float,2}`: n × m matrix of decision matrix. 
+ - `decisionMat::DataFrame`: n × m matrix of decision matrix in type of DataFrame. 
  - `weights::Array{Float64, 1}`: m-vector of weights for criteria.
  - `fs::Array{Function, 1}`: m-vector of functions that are either maximize or minimize for each single criterion.
  - `zeta::Float64`: zeta parameter for the algorithm. The default is 0.5.
