+++ docs/src/utility.md
@@ -1,0 +1,11 @@
+# Utility functions
+
+## mcdm
+```@docs
+JMcDM.mcdm
+```
+
+## summary
+```@docs
+JMcDM.summary
+```
