+++ paper/paper.md
@@ -20,7 +20,7 @@ affiliations:
    index: 1
  - name: Department of Transportation and Logistics, Istanbul University, Istanbul, Turkey
    index: 2
- - name: Department of Computer Engineering, Istanbul Technical University, Istanbul, Turkey
+ - name: Independent researcher
    index: 3
 
 
@@ -29,13 +29,13 @@ bibliography: paper.bib
 ---
 
 # Summary
-```JMcDM``` is a ```Julia``` package that implements some leading multiple-criteria decision making tools for both researchers and developers. ```Julia```'s REPL is well suited for researchers to perform their analysis using different methods and comparing their results. The package also provides the necessary infrastructure and utility functions for writing recently published methods.  The proposed package has brought MCDA tools to a relatively new language such as ```Julia``` with its significant performance promises. The methods developed in the package are also designed to be familiar to users who previously used ```R``` and ```Python``` languages. The paper presents the basics of the design, example usage, and code snippets.
+```JMcDM``` is a ```Julia``` package that implements some leading multiple-criteria decision making tools for both researchers and developers. ```Julia```'s REPL is well suited for researchers to perform their analysis using different methods and comparing their results. ```JMcDM``` also provides the necessary infrastructure and utility functions for writing recently published methods.  The proposed package has brought MCDM tools to a relatively new language such as ```Julia``` with its significant performance promises. The methods developed in the package are also designed to be familiar to users who previously used ```R``` and ```Python``` languages. The paper presents the basics of the design, example usage, and code snippets.
 
 
 # State of the field
 The one-dimensional array $a$ is in ascending order if and only if $a_i \le a_{i+1}$ where $i = 1, 2, \dots, n-1$, and $n$ is the length of array. In other terms, the process of ordering numbers requires the logical $\le$ operator to be perfectly defined. Since the operator $\le$ is not defined for any set of points in higher dimensions, $\mathbb{R}^p$ for $p \ge 2$, there is not a unique ordering of points.
 
-In multi-dimensional case, the binary domination operator $\succ$ applied on points $a$ and $b$, $a \succ b$, is true iif each item in $a$ is not worse than the correspoing item in $b$ and at least one item is better than the corresponding item in $b$ [@Deb_2002]. On the other hand, the more relaxed operator $\succeq$ returns true if each item in $a$ is as good as the corresponding item in $b$ [@greco2016multiple]. Several outranking methods in MCDA (Multiple-Criteria Decision Analysis) define a unique ranking mechanism to select the best alternative among others.
+In multi-dimensional case, the binary domination operator $\succ$ applied on points $a$ and $b$, $a \succ b$, is true iif each item in $a$ is not worse than the correspoing item in $b$ and at least one item is better than the corresponding item in $b$ [@Deb_2002]. On the other hand, the more relaxed operator $\succeq$ returns true if each item in $a$ is as good as the corresponding item in $b$ [@greco2016multiple]. Several outranking methods in MCDM (Multiple-Criteria Decision Making) define a unique ranking mechanism to select the best alternative among others.
 
[…]
