+++ README.md
@@ -121,3 +121,20 @@ julia> result.bestIndex
 2
 ```
 
+
+# Community guidelines
+
+Do you want to contribute?
+
+- Please create an Issue first. In this issue, please specify the idea.
+- Let the community discuss the new contribution.
+
+If the community decision is yes, please
+
+
+- Fork the repository
+- Send a pull request.
+
+Please read the issue [Welcome to newcomers!](https://github.com/jbytecode/JMcDM/issues/3) for implementation details.
+
+Welcome!
