+++ docs/src/index.md
@@ -2,6 +2,6 @@
 
 [The GitHub Repository](https://github.com/jbytecode/JMcDM)
 ```@contents
-Pages = ["index.md", "mcdms.md", "game.md"]
+Pages = ["index.md", "mcdms.md", "game.md", "dataenvelop.md", "utility.md"]
 Depth = 3
 ```
