### 2021-07-04 #3430 @jbytecode
Dear reviewer @sylvaticus

Thank you very much for your suggestions. Both the software and the paper are now better thanks to your comments. Here is the summary of my changes:

- Added `mcdm(df, w, fns, method)` type dispatch. For example

`julia> mcdm(df, w, fns, TopsisMethod()) `

now works. All of the methods are implemented in the same way and the methods that have optional parameters work similar like

`julia> mcdm(df, w, fns, XXXMethod(paramer, parameter))`

where `XXXMethod` inherits the type `MCDMMethod`.  These implementation covers the method specific arguments, for example, for Promethee it is like

`julia> mcdm(setting, PrometheeMethod(prefs, qs, ps))`

- `makeminmax()` added for this:
TOPSIS: Error During Test at test.jl:104
  Got exception outside of a @test

MethodError: no method matching topsis(::DataFrame, ::Vector{Float64}, ::Vector{typeof(maximum)})
  Closest candidates are:
    topsis(::DataFrame, ::Vector{Float64}, ::Vector{Function})

When we use [maximum, maximum, maximum], julia recognizes the vector in type of Vector{typeof(maximum)} rather than Function. When the function array includes both, no errors are thrown. I left it as is.

- v in vikor, zeta in grey etc. are now optional keyword arguments in a more standard way of implementation.

- Added `print(t <: MCDMResult)` functions for all MCDM methods.

- Definition of Promethee was fixed.

- MCDMSetting type is defined and `method(setting::MCDMSetting)` is implemented
for the methods that take df, w, fns as arguments. This type is documented as well. This is also current for the

`julia> mcdm(setting, TopsisMethod())`

type calls.

- documentation added for `summary()`. `summary(setting::MCDMSetting, methods)` type function call is also current for this.

- link to GitHub repo added in docs. New sections added to online docs.

- A note has been added to documentation about C and D values of Electre method.

- Dematel's input values are defined in the documentation.

- Documentation added for undocumented `makeDecisionMatrix()`. It is right that is converts a matrix to a DataFrame object, additionally add some nice column headers. this function is not vital but it is useful.

- A new documentation section is added for summary, mcdm, makeDecisionMatrix, etc.

- The article has been updated based on the comments. Statement of need section is now comparing the implemented package with the cited ones. Citations are ordered alphabetically. Typo fixed for multiple citations in Vikor. Citations are types in the standard JOSS template.

- Numerous additional test cases now covers the newly implemented interfaces for MCDMMethod, mcdm, MCDMSetting, etc.

- A descriptive "community guidelines" was also added in the repo.
