+++ README.md
@@ -1,5 +1,23 @@
 # JMcDM
-A package for Multiple-criteria decision making techniques in Julia
+A package for Multiple-criteria decision making techniques in Julia.
+
+# The problem
+
+Suppose a decision process has $n$ alternatives and $m$ criteria  which are either to be maximized or minimized. Each single criterion has a weight $0 <= w_i <= 1$ where sum of w_i is 1. f_i is either maximum or minimum. g_j(.) is evolution function and it is taken as g_j(x) = x in many methods. A multiple criteria decision problem can be represented using the decision table 
+
+   | **Criteria**  |   C_1    |   C_2    | ...  |   C_m    |
+   | :-----------: | :--------: | :--------: | :------: | :--------: |
+   |  **Weights**  |   w_1    |    w_2     | ...      |   w_m     |
+   | **Functions** |   f_1    |    f_2     | ...      |   f_m     |
+   |     A_1     | g_1(A_1)   |  g_2(A_1)  | ...      |  g_m(S_A)  |
+   |     A_2     | g_1(A_2)   |  g_2(A_2)  | ...      |  g_m(A_2)  |
+   |       ⋮       |     ⋮     |     ⋮      | ...      |     ⋮      |
+   |     A_n     | g_1(A_n)   |  g_2(A_n)  | ...      |  g_m(A_n)  |
+
+\noindent without loss of generality. When A_1, A_2, ..., A_n are alternatives and C_1, C_2, ..., C_m are different situations of a single criterion then the decision problem is said to be single criterion decision problem. If A_i and C_j are strategies of two game players then g_j(A_i) is the gain of the row player when she selects the strategy *i* and the column player selects the strategy C_j. 
+
+
+The package mainly focused on solving these kind of decision problems.
 
 
 ## Installation
