### 2021-07-06 bb37b46
add docs to aras method
