### 2021-07-06 6c91afd
add docs for electre
