### 2021-07-06 ca32116
add docs for vikor
