### 2021-07-06 e8f77f5
docs for MCDMSetting
