+++ src/types.jl
@@ -17,7 +17,14 @@ abstract type MCDMMethod end
 - `weights::Array{Float64,1}`: Array of weights for each criterion.
 - `fns::Array{Function, 1}`: Array of functions. The elements are either minimum or maximum.
 
+# Description 
+Many methods including Topsis, Electre, Waspas, etc., use a decision matrix, weights, and directions
+of optimizations in types of DataFrame, Vector, and Vector, respectively. The type MCDMSetting simply
+holds these information to pass them into methods easly. Once a MCDMSetting object is created, the problem
+can be passed into several methods like topsis(setting), electre(setting), waspas(setting), etc.  
+
 # Examples
+
 ```julia-repl
 julia> df = DataFrame();
 julia> df[:, :x] = Float64[9, 8, 7];
@@ -34,6 +41,8 @@ julia> setting = MCDMSetting(df, w, fns)
 julia> result = topsis(setting);
 julia> # Same result can be obtained using
 julia> result2 = mcdm(setting, TopsisMethod())
+```
+
 """
 struct MCDMSetting 
     df::DataFrame
