### 2021-07-06 af124e0
add MCDMSetting documentary in offline docs
