+++ docs/src/utility.md
@@ -1,5 +1,10 @@
 # Utility functions
 
+## MCDMSetting
+```@docs
+JMcDM.MCDMSetting
+```
+
 ## mcdm
 ```@docs
 JMcDM.mcdm
