### 2021-07-06 #3430 @jbytecode
Dear reviewer @sylvaticus

Recent changes has been pumped in v0.2.0. Today, I have just released a new version of v0.2.1 with new documentation and several changes/corrections in both repo, online docs, and source code. The new version is ready to update in Julia.

Thank you.
