### 2021-07-13 #3430 @sylvaticus
Hello, thank you for your quick reaction. I tested the changes you made on the readme, online documentation, JOSS paper and the software and that I believe have improved the software and its user experience.

The two main points (scope of the software and methods integration) that I have highlighted have been addressed respectively in the readme page and in the new "high-level" functions.

There are still a couple of very minor errors present that I may have been bad to communicate in my review (English is not my mother tongue):

- `electre` function documentation: the output is copy/pasted from the `topsis` function above, but it has not been adapted. The `electre` documentation present the output as a "TopsisResult object" while it is a "ElectreResult" object
- The first parameter of `AHP` function is an array of DataFrames, but in the documentation it is presented as a single DataFrame
- I appreciated the added documentation in the `damatel` function, but still I don't get why you use `n x m` matrix... it seems to me that the input matrix of the function should be squared, so `n x n` ??

There remains some of the  "nice to have" things that I have suggested, but that of course do not preclude in my opinion the publication of the JOSS paper, like the coverage, the documentation divided by releases and a documentation that accompany each releases. I have noticed indeed that you don't have releases notes. You can easily  add release notes in the same JuliaRegistrator command:

```
@JuliaRegistrator register

Release notes:

    - bla bla bla
    - bla bla bla

```

Finally, I didn't understand your reply "Statement of need section is now comparing the implemented package with the cited ones.". This was already the case for the first version of the paper. It is my understanding that the intended section for comparison with existing software would be instead the "State of the field" section, with the "Statement of need" section more describing  the problem to which the software provides solutions (more or less what it is now in "State of the field"). However, I don't know if this is strictly enforced  at JOSS, or if it is just a recommendation.
