### 2021-07-13 d52666f
corrrect docstring for electre
