### 2021-07-14 #3430 @jbytecode
- [X] `electre` function documentation: the output is copy/pasted from the `topsis` function above, but it has not been adapted. The `electre` documentation present the output as a "TopsisResult object" while it is a "ElectreResult" object
- [X] The first parameter of `AHP` function is an array of DataFrames, but in the documentation it is presented as a single DataFrame
- [X] I appreciated the added documentation in the `damatel` function, but still I don't get why you use `n x m` matrix... it seems to me that the input matrix of the function should be squared, so `n x n` ??
- [X] The manuscript sections are refactored as recommended above.
