### 2021-07-16 #3430 @brunaw
Dear @jbytecode,

Here's my review/comments abou the submission, please let me know if you have any questions about it:

## General

- The summary section seem to have a few disconnected sentences. Could the authors make the text flow better?

- The `State of the field` section of the paper needs some changes. The first paragraphs are not a literature review per se, but rather an introduction/explanation of the method. That could probably go in a "Introduction" or "Methods" section, and then the actual LR paragraph could be the "Literature review" section. Does that make sense?

- The last paragraph of the `State of the field` needs at least some connections between the sentences. The way it's currently written make it look like just a list of software, with no comments or details, so it's hard to understand why are they relevant to this submission

- In the "Statement of need" section, there is very little written
about the actual need of the package. My suggestion is that the
authors complement that.

- In the "Community guidelines", what does:

> "Let the community discuss the new contribution"

mean in practice? Could the authors please specify that in the text? I understand the idea but the way that's written seem a bit confusing to whomever might want to contribute

- There are no guidelines on error reporting/seeking support,  which could be added to the Community guidelines

## Package content/documentation

  - I haven't found many usage examples for the functions (besides in the documentation). Do the authors plan to add that, with e.g. a package manual?
  - There is no list of dependencies explicitly written for the users to see
  - It would be nice to say why you are also loading `DataFrames` in the code example

## Package code
- The package installs locally with no issues, but the precompilation takes a while (at least on my computer). Is that normal?

Bruna
