+++ CONTRIBUTING.md
@@ -1,15 +1,16 @@
 Do you want to contribute?
 
 - Please create an Issue first. In this issue, please specify the idea.
-- Let the community discuss the new contribution.
+- Let the community discuss the new contribution in our Slack channel or in the created issue.
 
 If the community decision is yes, please
-
 
 - Fork the repository
 - Send a pull request.
 
 Please read the issue [Welcome to newcomers!](https://github.com/jbytecode/JMcDM/issues/3) for implementation details.
+
+Our Slack channel is [JMcDM Slack Channel](https://julialang.slack.com/archives/C01MJ0VF1U3).
 
 Welcome!
 
