### 2021-07-17 5518eed
add docs for Laplace
