### 2021-07-17 9647c74
add docs for Maximin
