### 2021-07-17 7656909
add docs for Minimax
