+++ src/singlecriterion.jl
@@ -95,8 +95,29 @@ end
 
 
 """
-    Apply maximax method for a given decision matrix
+    maximax(decisionMat)
 
+    Apply Maximax method for a given decision matrix (for convenience, in type of DataFrame).
+
+# Arguments:
+ - `decisionMat::DataFrame`: Decision matrix with n alternatives and m criteria. 
+ 
+# Output 
+- `::MaximaxResult`: MaximaxResult object that holds multiple outputs including the best alternative.
+
+# Examples
+```julia-repl
+julia> mat = [
+        26 26 18 22;
+        22 34 30 18;
+        28 24 34 26;
+        22 30 28 20
+    ]
+
+julia> dm = makeDecisionMatrix(mat)
+
+julia> result = maximax(dm)
+```
 """
 function maximax(decisionMatrix::DataFrame)::MaximaxResult
 
@@ -119,8 +140,29 @@ end
 
 
 """
-    Apply minimax method for a given decision matrix
+    minimax(decisionMat)
 
+    Apply Minimax method for a given decision matrix (for convenience, in type of DataFrame).
+
+# Arguments:
+ - `decisionMat::DataFrame`: Decision matrix with n alternatives and m criteria. 
+ 
+# Output 
+- `::MinimaxResult`: MinimaxResult object that holds multiple outputs including the best alternative.
+
+# Examples
+```julia-repl
+julia> mat = [
+        26 26 18 22;
+        22 34 30 18;
+        28 24 34 26;
+        22 30 28 20
+    ]
+
+julia> dm = makeDecisionMatrix(mat)
+
+julia> result = minimax(dm)
+```
 """
 function minimax(decisionMatrix::DataFrame)::MinimaxResult
 
