### 2021-07-17 09beed6
add docs for single criterion tools
