+++ docs/src/scdm.md
@@ -1,0 +1,51 @@
+# Single-criterion decision making tools
+
+## laplace
+```@docs
+JMcDM.laplace
+```
+
+## maximin
+```@docs
+JMcDM.maximin
+```
+
+## maximax
+```@docs
+JMcDM.maximax
+```
+
+
+## minimax
+```@docs
+JMcDM.minimax
+```
+
+## minimin
+```@docs
+JMcDM.minimin
+```
+
+## savage
+```@docs
+JMcDM.savage
+```
+
+## hurwicz
+```@docs
+JMcDM.hurwicz
+```
+
+## mle
+```@docs
+JMcDM.mle
+```
+
+
+## expectedregret
+```@docs
+JMcDM.expectedregret
+```
+
+
+
