### 2021-07-17 #3430 @jbytecode
Dear reviewer @brunaw,

Firstly, thank you for your valuable comments. Thanks to your suggestions, the new version of the package is now better. Please let me know if some of the listed issues are not complete or if I unluckily misunderstood something. Here is a list of my changes:

> Dear @jbytecode,
>
> Here's my review/comments abou the submission, please let me know if you have any questions about it:
>
> ## General
> * The summary section seem to have a few disconnected sentences. Could the authors make the text flow better?

- [X] The summary section is refactored.

> * The `State of the field` section of the paper needs some changes. The first paragraphs are not a literature review per se, but rather an introduction/explanation of the method. That could probably go in a "Introduction" or "Methods" section, and then the actual LR paragraph could be the "Literature review" section. Does that make sense?

- [X] State of the field is divided into an Introduction and a State of the field section. The former basically describes the problem whereas the latter is focused on the current developments on the subject.

> * The last paragraph of the `State of the field` needs at least some connections between the sentences. The way it's currently written make it look like just a list of software, with no comments or details, so it's hard to understand why are they relevant to this submission

- [X] The section is refactored.

> * In the "Statement of need" section, there is very little written
>   about the actual need of the package. My suggestion is that the
>   authors complement that.

- [X] This issue is now more clear.

> * In the "Community guidelines", what does:
>
> > "Let the community discuss the new contribution"
>
> mean in practice? Could the authors please specify that in the text? I understand the idea but the way that's written seem a bit confusing to whomever might want to contribute
>
> * There are no guidelines on error reporting/seeking support,  which could be added to the Community guidelines
>
- [X] Clarified. Added our slack channel for discussions and possible conversations.
-
> ## Package content/documentation
> * I haven't found many usage examples for the functions (besides in the documentation). Do the authors plan to add that, with e.g. a package manual?

- [X] Added new documentation (and examples) for single-criterion methods such as minimax, maximin, savage, maximum likelihood, etc. A new version of the package has been pumped to Julia repository. Online documentary has been updated.
-
> * There is no list of dependencies explicitly written for the users to see

- [X] Since the Julia package manager automatically downloads and installs the dependencies, a standard user is not interested in this list. However, a list of dependencies added to README.md file.

> * It would be nice to say why you are also loading `DataFrames` in the code example

- [X] The package now exports DataFrame, so there is no longer need to require this. The examples in both in paper and repo are updated.
-
>
> ## Package code
> * The package installs locally with no issues, but the precompilation takes a while (at least on my computer). Is that normal?

- [ ] Since the package is dependent on the some heavy-weight packages such like JuMP and Cbc, it is normal. This issue is current for many Julia package just because the Julia first pre-compiles the whole package and its dependencies before the first use. Nothing to do for this (at least as I know).
-
>
> Bruna
