+++ src/print.jl
@@ -1,154 +1,154 @@
-function print(result::TopsisResult)
-    println("Scores:")
-    println(result.scores)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::TopsisResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::ARASResult)
-    println("Scores:")
-    println(result.scores)
-    println("Orderings: ")
-    println(result.orderings)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::ARASResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Orderings: ")
+    println(io, result.orderings)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::CODASResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ranking: ")
-    println(result.ranking)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::CODASResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Ranking: ")
+    println(io, result.ranking)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
 
-function print(result::COPRASResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ranking: ")
-    println(result.ranking)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::COPRASResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Ranking: ")
+    println(io, result.ranking)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::CRITICResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ranking: ")
-    println(result.ranking)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::CRITICResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Ranking: ")
+    println(io, result.ranking)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::CoCoSoResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ranking: ")
-    println(result.ranking)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::CoCoSoResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Ranking: ")
+    println(io, result.ranking)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::EDASResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ranking: ")
-    println(result.ranking)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::EDASResult)
+    println(io, "Scores:")
+    println(io, result.scores)
+    println(io, "Ranking: ")
+    println(io, result.ranking)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::ElectreResult)
-    println("Best indices:")
-    println(result.bestIndex)
+function Base.show(io::IO, result::ElectreResult)
+    println(io, "Best indices:")
+    println(io, result.bestIndex)
 end
 
-function print(result::GreyResult)
-    println("Scores:")
-    println(result.scores)
-    println("Ordering: ")
-    println(result.ordering)
[…]
