+++ CHANGELOG.md
@@ -1,0 +1,2 @@
+### the candidate release
+- Base.show(io:IO, MCDMResult) implementations for pretty printing all of the results
