+++ README.md
@@ -5,6 +5,9 @@ A package for Multiple-criteria decision making techniques in Julia.
 
 Suppose a decision process has n alternatives and m criteria  which are either to be maximized or minimized. Each single criterion has a weight <img src="https://render.githubusercontent.com/render/math?math=0 \le w_i \le 1"> where sum of <img src="https://render.githubusercontent.com/render/math?math=w_i"> is 1. <img src="https://render.githubusercontent.com/render/math?math=f_i"> is either maximum or minimum. <img src="https://render.githubusercontent.com/render/math?math=g_j(.)"> is evolution function and it is taken as <img src="https://render.githubusercontent.com/render/math?math=g_j(x)=x"> in many methods. A multiple criteria decision problem can be represented using the decision table 
 
+<img src="https://github.com/jbytecode/JMcDM/blob/gh-pages/images/generalformula.png" width = "50%"/>
+
+<!--
    | **Criteria**  |   C_1    |   C_2    |  ...  |   C_m    |
    | :-----------: | :------: | :------: | :---: | :------: |
    |  **Weights**  |   w_1    |   w_2    |  ...  |   w_m    |
@@ -13,6 +16,7 @@ Suppose a decision process has n alternatives and m criteria  which are either t
    |      A_2      | g_1(A_2) | g_2(A_2) |  ...  | g_m(A_2) |
    |       ⋮       |    ⋮     |    ⋮     |  ...  |    ⋮     |
    |      A_n      | g_1(A_n) | g_2(A_n) |  ...  | g_m(A_n) |
+-->
 
 without loss of generality. When <img src="https://render.githubusercontent.com/render/math?math=A_1, A_2, \ldots, A_n"> are alternatives and <img src="https://render.githubusercontent.com/render/math?math=C_1, C_2, \ldots, C_m"> are different situations of a single criterion then the decision problem is said to be single criterion decision problem. If <img src="https://render.githubusercontent.com/render/math?math=A_i"> and <img src="https://render.githubusercontent.com/render/math?math=C_j"> are strategies of two game players then <img src="https://render.githubusercontent.com/render/math?math=g_j(A_i)"> is the gain of the row player when she selects the strategy *i* and the column player selects the strategy <img src="https://render.githubusercontent.com/render/math?math=C_j">. 
 
