### 2021-09-04 f62f0ea
import (only) needed functions from external libraries
