+++ test/runtests.jl
@@ -1,5 +1,5 @@
 using Test
-using DataFrames
+import DataFrames: DataFrame, DataFrameRow
 
 
 using JMcDM
@@ -37,6 +37,7 @@ using JMcDM
     end
 end
 
+@testset "Utility functions" begin
 @testset "Euclidean distance" begin
     @test euclidean([0.0, 1.0, 2.0], [0.0, 1.0, 2.0]) == 0.0
     @test euclidean([0.0, 0.0, 0.0]) == 0.0
@@ -132,6 +133,383 @@ end
     @test names(dfwithnames)[4] == "EE"
     @test names(dfwithnames)[5] == "FG"
 end
+
+
+@testset "Reverse minimum & maximum array" begin
+
+    fns = [minimum, maximum, maximum, minimum, maximum]
+    revfns = [maximum, minimum, minimum, maximum, minimum]
+
+    @test reverseminmax(fns) == revfns 
+
+    @test reverseminmax(revfns) == fns 
+end
+
+@testset "Make Array of minimum and maximum" begin
+    result1 = makeminmax([maximum, maximum, maximum, maximum])
+
+    @test typeof(result1) == Array{Function,1}
+
+    @test typeof(result1[1]([1.0, 2.0, 3.0])) == Float64 
+
+    @test result1[1]([1.0, 2.0, 3.0]) == 3.0 
+
+end
+
+end # end of utility function test
+
+
+@testset "Single Criterion Decision Making tools" begin
+@testset "Laplace" begin
+    
+    tol = 0.00001
+
+    mat = [
+        3000 2750 2500 2250;
+        1500 4750 8000 7750;
+        2000 5250 8500 11750
+    ]
+
+    dm = makeDecisionMatrix(mat)
+
+    result = laplace(dm)
+
+    @test isa(result, LaplaceResult)
+
+    @test isapprox(result.expected_values, [2625.0, 5500, 6875], atol=tol)
+
+    @test result.bestIndex == 3
+end
+
+
+@testset "Maximin" begin
+    
+    tol = 0.00001
+
+    mat = [
+        26 26 18 22;
+        22 34 30 18;
+        28 24 34 26;
+        22 30 28 20
+    ]
+
+    dm = makeDecisionMatrix(mat)
+
+    result = maximin(dm)
+
+    @test isa(result, MaximinResult)
+
+    @test isapprox(result.rowmins, [18, 18, 24, 20], atol=tol) 
+
+    @test result.bestIndex == 3
+end
+
+
+@testset "Maximax" begin
+    
+    tol = 0.00001
+
+    mat = [
+        26 26 18 22;
+        22 34 30 18;
+        28 24 34 26;
+        22 30 28 20
+    ]
+
+    dm = makeDecisionMatrix(mat)
+
+    result = maximax(dm)
+
+    @test isa(result, MaximaxResult)
+
+    @test isapprox(result.rowmaxs, [26, 34, 34, 30], atol=tol) 
+     
+    @test result.bestIndex in [2, 3]
+end
+
+
+@testset "Minimax" begin
+    
+    tol = 0.00001
+
+    mat = [
+        26 26 18 22;
+        22 34 30 18;
+        28 24 34 26;
+        22 30 28 20
+    ]
+
+    dm = makeDecisionMatrix(mat)
+
+    result = minimax(dm)
+
+    @test isa(result, MinimaxResult)
+
+    @test isapprox(result.rowmaxs, [26, 34, 34, 30], atol=tol) 
+     
+    @test result.bestIndex == 1
+end
+
+
+@testset "Minimin" begin
+    
+    tol = 0.00001
+
+    mat = [
+        26 26 18 22;
[…]
