### 2021-09-04 8f89810
removed StatsBase from dependencies and geomean implemented
