+++ src/JMcDM.jl
@@ -6,13 +6,11 @@ import LinearAlgebra: I
 import JuMP: @variable, @objective, @constraint
 import JuMP: Model, MOI, optimize!, JuMP, objective_value
 import Cbc: Cbc
-import StatsBase: geomean
 import Statistics: cor, mean, std
 
 # for Pretty printing
 # of MCDM results
 import Base.show
-
 
 # includes 
 include("types.jl")
