### 2021-09-04 41a45bc
removed dependency for Statistics, cor() implemented
