+++ src/JMcDM.jl
@@ -6,7 +6,6 @@ import LinearAlgebra: I
 import JuMP: @variable, @objective, @constraint
 import JuMP: Model, MOI, optimize!, JuMP, objective_value
 import Cbc: Cbc
-import Statistics: cor, mean, std
 
 # for Pretty printing
 # of MCDM results
