### 2021-09-04 b6ccbe8
removed dependency for LinearAlgebra, I() implemented
