+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-### the candidate release
+### 0.2.4
 - Base.show(io:IO, MCDMResult) implementations for pretty printing all of the results
+- I() implemented, LinearAlgebra package removed from dependencies.
+- mean(), geomean(), std(), and cor() are implemented and StatsBase & Statistics packages are removed
+- using keyword replaced by import and only the needed functions are loaded at startup.
+- 
