### 2021-09-04 #3430 @jbytecode
Dear reviewer @brunaw

Here is the minor changes for v0.2.4.

- Base.show(io:IO, <: MCDMResult) implementations for pretty printing all of the results instead of overloading print().
- I() implemented, LinearAlgebra package removed from dependencies.
- mean(), geomean(), std(), and cor() are implemented and StatsBase & Statistics packages are removed from dependencies.
- `using` keyword replaced by `import` and only the needed functions are explicitly declared for requirements.
- Dependency list is updated.

I know these are standard packages and do not bring a substantial overhead but I think the first startup of the package is a little bit faster now (this is not a performance claim of course).
