+++ src/dataenvelop.jl
@@ -64,7 +64,6 @@ function dataenvelop(input::Array{Float64,2}, output::Array{Float64,1}; verbose:
 
         model = Model(Cbc.Optimizer);
         MOI.set(model, MOI.Silent(), !verbose)
-
         @variable(model, x[1:nrow])
         @variable(model, theta)
 
