+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 0.2.5
+- On/Off switch for tests. 
+
 ### 0.2.4
 - Base.show(io:IO, MCDMResult) implementations for pretty printing all of the results
 - I() implemented, LinearAlgebra package removed from dependencies.
