+++ test/runtests.jl
@@ -11,7 +11,7 @@ const testLPBasedFunctions = true
 
 
 if testUtilityFunctions
-@info "Testing Utility Functions"
+@info "Testing Utility and Printing Functions"
 @testset "Pretty printing of results" begin
     @testset "Pretty printing of TopsisResult" begin
         t = TopsisResult(
@@ -45,6 +45,46 @@ if testUtilityFunctions
 end
 
 @testset "Utility functions" begin
+    @testset "Identity matrix" begin
+        @test JMcDM.I(2) == [1.0 0; 0 1]
+        @test JMcDM.I(3) == [1.0 0 0; 0 1 0; 0 0 1]
+        @test JMcDM.I(4) == [1.0 0 0 0; 0 1 0 0; 0 0 1 0; 0 0 0 1]
+    end
+
+    @testset "mean" begin
+        @test JMcDM.mean([1.0, 2.0, 3.0]) == 2.0
+        @test JMcDM.mean([1.0, 2.0, 3.0, 4.0]) == 2.5
+    end
+
+    @testset "var" begin
+        @test JMcDM.var([1.0, 2.0, 3.0]) == 1.0
+        @test isapprox(JMcDM.var([1.0, 2.0, 3.0, 4.0]), 1.666667, atol = 0.001)
+    end
+
+    @testset "std" begin
+        @test JMcDM.std([1.0, 2.0, 3.0]) == 1.0
+        @test isapprox(JMcDM.std([1.0, 2.0, 3.0, 4.0]), 1.290994, atol = 0.001)
+    end
+
+    @testset "geomean" begin
+        @test isapprox(JMcDM.geomean([1.0, 2.0, 3.0]), 1.817121, atol = 0.001)
+        @test isapprox(JMcDM.geomean([1.0, 2.0, 3.0, 4.0]), 2.213364, atol = 0.001)
+    end
+
+    @testset "cor" begin
+        x = [1.0, 2.0, 3.0, 4.0, 5.0]
+        y = [5.0, 4.0, 3.0, 2.0, 1.0]
+        @test JMcDM.cor(x, x) == 1.0
+        @test JMcDM.cor(x, y) == -1.0
+
+        mat = hcat(x, y)
+        cormat = JMcDM.cor(mat)
+        @test cormat[1, 1] == 1.0
+        @test cormat[1, 2] == -1.0
+        @test cormat[2, 1] == -1.0
+        @test cormat[2, 2] == 1.0 
+    end
+
 @testset "Euclidean distance" begin
     @test euclidean([0.0, 1.0, 2.0], [0.0, 1.0, 2.0]) == 0.0
     @test euclidean([0.0, 0.0, 0.0]) == 0.0
