+++ src/.dev/test.jl
@@ -1,1293 +1,0 @@
-using Base: Float64
-using Test
-import DataFrames: DataFrame
-
-
-include("playground.jl")
-
-@testset "Pretty printing of results" begin
-    @testset "Pretty printing of TopsisResult" begin
-        t = TopsisResult(
-            DataFrame(rand(3,4), :auto),
-            rand(3),
-            DataFrame(rand(3,4), :auto),
-            DataFrame(rand(3,4), :auto),
-            1,
-            [1, 2, 3])
-        io = IOBuffer()
-        show(io, t)
-        str_expected = String(take!(io))
-        @test str_expected == "Scores:\n[1.0, 2.0, 3.0]\nBest indices:\n1\n"
-    end
-
-    @testset "Pretty printing of ARAS Result" begin
-        t = ARASResult(
-            rand(3),
-            rand(3,3),
-            rand(3,3),
-            rand(3),
-            [1.0, 2.0, 3.0],
-            [3, 2, 1],
-            2
-        )
-        io = IOBuffer()
-        show(io, t)
-        str_expected = String(take!(io))
-        @test str_expected == "Scores:\n[1.0, 2.0, 3.0]\nOrderings: \n[3, 2, 1]\nBest indices:\n2\n" 
-    end
-end
-
-@testset "Euclidean distance" begin
-    @test euclidean([0.0, 1.0, 2.0], [0.0, 1.0, 2.0]) == 0.0
-    @test euclidean([0.0, 0.0, 0.0]) == 0.0
-    @test euclidean([0.0, 0.0, 1.0]) == 1.0
-    @test euclidean([0.0, 0.0, 1.0], [0.0, 0.0, 2.0]) == 1.0
-    @test euclidean([0, 0, 0]) == 0.0
-    @test euclidean([0, 0, 0, 0], [1, 0, 0, 0]) == 1.0
-    @test euclidean([0, 0, 0, 0], [1, 0, 0, 0]) isa Float64
-    @test euclidean([0, 0, 0, 0], [1.0, 0, 0, 0]) isa Float64
-end
-
-@testset "Normalization" begin
-    tol = 0.00001
-    nz = normalize([1.0, 2.0, 3.0, -1.0, 0.0])
-    @test isapprox(nz[1], 0.2581989, atol=tol)
-    @test isapprox(nz[2], 0.5163978, atol=tol)
-    @test isapprox(nz[3], 0.7745967, atol=tol)
-    @test isapprox(nz[4], -0.2581989, atol=tol)
-    @test isapprox(nz[5], 0.0000000, atol=tol)
-
-    nzint = normalize([1, 2, 3, -1, 0])
-    @test isapprox(nzint[1], 0.2581989, atol=tol)
-    @test isapprox(nzint[2], 0.5163978, atol=tol)
-    @test isapprox(nzint[3], 0.7745967, atol=tol)
-    @test isapprox(nzint[4], -0.2581989, atol=tol)
-    @test isapprox(nzint[5], 0.0000000, atol=tol)
-end
-
-@testset "Column min and max vectors" begin
-    df = DataFrame()
-    df[:,:x] = [0.0, 1.0, 10.0]
-    df[:,:y] = [0.0, -1.0, -10.0]
-    @test colmins(df) == [0.0, -10.0]
-    @test colmaxs(df) == [10.0, 0.0]
-end
-
-@testset "Unitize vector" begin
-    x = [1.0, 2.0, 3.0, 4.0, 5.0]
-    result = x |> unitize |> sum
-    @test result == 1.0
-
-    x = [1, 2, 3, 4, 5]
-    result = x |> unitize |> sum 
[…]
