+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '^1.6.0-0'
+          - '1.6.2'
         os:
           - ubuntu-latest
           - macOS-latest
