+++ test/testutility.jl
@@ -1,0 +1,195 @@
+@testset "Utility and Printing Functions" begin
+    @testset "Pretty printing of results" begin
+        @testset "Pretty printing of TopsisResult" begin
+            t = TopsisResult(
+                DataFrame(rand(3, 4), :auto),
+                rand(3),
+                DataFrame(rand(3, 4), :auto),
+                DataFrame(rand(3, 4), :auto),
+                1,
+                [1, 2, 3],
+            )
+            io = IOBuffer()
+            show(io, t)
+            str_expected = String(take!(io))
+            @test str_expected == "Scores:\n[1.0, 2.0, 3.0]\nBest indices:\n1\n"
+        end
+
+        @testset "Pretty printing of ARAS Result" begin
+            t = ARASResult(
+                rand(3),
+                rand(3, 3),
+                rand(3, 3),
+                rand(3),
+                [1.0, 2.0, 3.0],
+                [3, 2, 1],
+                2,
+            )
+            io = IOBuffer()
+            show(io, t)
+            str_expected = String(take!(io))
+            @test str_expected ==
+                  "Scores:\n[1.0, 2.0, 3.0]\nOrderings: \n[3, 2, 1]\nBest indices:\n2\n"
+        end
+    end
+
+    @testset "Utility functions" begin
+        @testset "Identity matrix" begin
+            @test JMcDM.I(2) == [1.0 0; 0 1]
+            @test JMcDM.I(3) == [1.0 0 0; 0 1 0; 0 0 1]
+            @test JMcDM.I(4) == [1.0 0 0 0; 0 1 0 0; 0 0 1 0; 0 0 0 1]
+        end
+
+        @testset "mean" begin
+            @test JMcDM.mean([1.0, 2.0, 3.0]) == 2.0
+            @test JMcDM.mean([1.0, 2.0, 3.0, 4.0]) == 2.5
+        end
+
+        @testset "var" begin
+            @test JMcDM.var([1.0, 2.0, 3.0]) == 1.0
+            @test isapprox(JMcDM.var([1.0, 2.0, 3.0, 4.0]), 1.666667, atol = 0.001)
+        end
+
+        @testset "std" begin
+            @test JMcDM.std([1.0, 2.0, 3.0]) == 1.0
+            @test isapprox(JMcDM.std([1.0, 2.0, 3.0, 4.0]), 1.290994, atol = 0.001)
+        end
+
+        @testset "geomean" begin
+            @test isapprox(JMcDM.geomean([1.0, 2.0, 3.0]), 1.817121, atol = 0.001)
+            @test isapprox(JMcDM.geomean([1.0, 2.0, 3.0, 4.0]), 2.213364, atol = 0.001)
+        end
+
+        @testset "cor" begin
+            x = [1.0, 2.0, 3.0, 4.0, 5.0]
+            y = [5.0, 4.0, 3.0, 2.0, 1.0]
+            @test JMcDM.cor(x, x) == 1.0
+            @test JMcDM.cor(x, y) == -1.0
+
+            mat = hcat(x, y)
+            cormat = JMcDM.cor(mat)
+            @test cormat[1, 1] == 1.0
+            @test cormat[1, 2] == -1.0
+            @test cormat[2, 1] == -1.0
+            @test cormat[2, 2] == 1.0
+        end
+
+        @testset "Euclidean distance" begin
+            @test euclidean([0.0, 1.0, 2.0], [0.0, 1.0, 2.0]) == 0.0
+            @test euclidean([0.0, 0.0, 0.0]) == 0.0
+            @test euclidean([0.0, 0.0, 1.0]) == 1.0
+            @test euclidean([0.0, 0.0, 1.0], [0.0, 0.0, 2.0]) == 1.0
+            @test euclidean([0, 0, 0]) == 0.0
[…]
