+++ Project.toml
@@ -9,7 +9,7 @@ DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 
 [compat]
-Cbc = "^0"
+Cbc = "^0.8"
 DataFrames = "^1"
-JuMP = "^0"
+JuMP = "^0.21"
 julia = "^1.4"
