### 2021-09-20 #3430 @brunaw
Dear @jbytecode,

Thank you for the changes and my apologies for the delay in replying, indeed I wasn't getting this issue's notifications but I was reminded by email that I should finish this review.

I'm happy to mark all the boxes now, though I still feel like some of the writing could be improved. For example, in the documentation make sure that all 3 community guidelines topics (e.g. how to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support) are covered. The paper could also use a quick text review, maybe with a writing helper software like Grammarly (which always helps me a lot 😃 ).

Please let me know if you have any questions,

Best regards,
Bruna
