### 2021-09-21 #3430 @jbytecode
Dear reviewer @brunaw

The `README.md` file is updated.
- Contribute to the software, report issues, and seek support sections in `README.md` are refactored.
- The Contribute to the software part now includes new subsections: `Where to start` and `The design pattern`. In this subsections, we give a brief description of implementing a new method with minimum requirements.
- Typo and language are checked and corrected.

Here is the latest changes:
- https://github.com/jbytecode/JMcDM/commit/9a4ac6d8c30237958c8f6a70b1b765df8a58b29a
- https://github.com/jbytecode/JMcDM/commit/fe4416873842c21b2bf883602b2bf9b18e0c1171

Are those stuff okay for you?

Of course, now we are reading the whole paper, and we will check the whole stuff. Thank you for your response and valuable reviewing.
