### 2021-09-21 e6a5b2f
grammar check for paper.md
