+++ paper/paper.md
@@ -1,27 +1,27 @@
 ---
 title: 'JMcDM: A Julia package for multiple-criteria decision-making tools'
 tags:
-  - Julia
-  - decision making
-  - multiple criteria
-  - outranking
+  - Julia
+  - decision making
+  - multiple criteria
+  - outranking
 authors:
-  - name: Mehmet Hakan Satman
-    orcid: 0000-0002-9402-1982
-    affiliation: 1
-  - name: Bahadır Fatih Yıldırım
-    orcid: 0000-0002-0475-741X
-    affiliation: 2
-  - name: Ersagun Kuruca
-    orcid: 0000-0002-2552-7701
-    affiliation: 3
+  - name: Mehmet Hakan Satman
+    orcid: 0000-0002-9402-1982
+    affiliation: 1
+  - name: Bahadır Fatih Yıldırım
+    orcid: 0000-0002-0475-741X
+    affiliation: 2
+  - name: Ersagun Kuruca
+    orcid: 0000-0002-2552-7701
+    affiliation: 3
 affiliations:
- - name: Department of Econometrics, Istanbul University, Istanbul, Turkey
-   index: 1
- - name: Department of Transportation and Logistics, Istanbul University, Istanbul, Turkey
-   index: 2
- - name: Independent researcher
-   index: 3
+ - name: Department of Econometrics, Istanbul University, Istanbul, Turkey
+   index: 1
+ - name: Department of Transportation and Logistics, Istanbul University, Istanbul, Turkey
+   index: 2
+ - name: Independent researcher
+   index: 3
 
 
 date: 8 May 2021
@@ -29,37 +29,37 @@ bibliography: paper.bib
 ---
 
 # Summary
-```JMcDM``` is a ```Julia``` package that implements some leading multiple-criteria decision-making tools for both researchers and developers. By having a REPL tool, ```Julia``` is well suited for researchers to perform their analysis using different methods and comparing their results. ```JMcDM``` also provides the necessary infrastructure, utility functions, and a standardized API for implementing recently published methods.  The proposed package has brought MCDM (Multiple-Criteria Decision-Making) tools to a relatively new language such as ```Julia``` with its significant performance promises. Besides Julia being a new language, the methods developed in the package are designed to be familiar to users who previously used ```R``` and ```Python``` languages. The paper presents the basics of the design, example usage, and code snippets.
+```JMcDM``` is a ```Julia``` package that implements some leading multiple-criteria decision-making tools for both researchers and developers. By having a REPL tool, ```Julia``` is well suited for researchers to perform their analysis using different methods and comparing their results. ```JMcDM``` also provides the necessary infrastructure, utility functions, and a standardized API for implementing recently published methods.  The proposed package has brought MCDM (Multiple-Criteria Decision-Making) tools to a relatively new language such as ```Julia``` with its significant performance promises. Besides Julia being a new language, the methods developed in the package are designed to be familiar to users who previously used ```R``` and ```Python``` languages. The paper presents the basics of the design, example usage, and code snippets.
 
 
 # Introduction
 
[…]
