### 2021-09-21 3de164e
grammar check for paper.md
