### 2021-09-21 cad174b
grammar check for paper.md
