### 2021-09-21 d0450fa
grammar check for paper.md
