+++ README.md
@@ -1,5 +1,5 @@
 # JMcDM
-A package for Multiple-criteria decision making techniques in Julia.
+A package for Multiple-criteria decision-making techniques in Julia.
 
 ## The problem
 
@@ -18,14 +18,14 @@ Suppose a decision process has n alternatives and m criteria  which are either t
    |      A_n      | g_1(A_n) | g_2(A_n) |  ...  | g_m(A_n) |
 -->
 
-without loss of generality. When <img src="https://render.githubusercontent.com/render/math?math=A_1, A_2, \ldots, A_n"> are alternatives and <img src="https://render.githubusercontent.com/render/math?math=C_1, C_2, \ldots, C_m"> are different situations of a single criterion then the decision problem is said to be single criterion decision problem. If <img src="https://render.githubusercontent.com/render/math?math=A_i"> and <img src="https://render.githubusercontent.com/render/math?math=C_j"> are strategies of two game players then <img src="https://render.githubusercontent.com/render/math?math=g_j(A_i)"> is the gain of the row player when she selects the strategy *i* and the column player selects the strategy <img src="https://render.githubusercontent.com/render/math?math=C_j">. 
+without loss of generality. When <img src="https://render.githubusercontent.com/render/math?math=A_1, A_2, \ldots, A_n"> are alternatives and <img src="https://render.githubusercontent.com/render/math?math=C_1, C_2, \ldots, C_m"> are different situations of a single criterion then the decision problem is said to be single criterion decision problem. If <img src="https://render.githubusercontent.com/render/math?math=A_i"> and <img src="https://render.githubusercontent.com/render/math?math=C_j"> are strategies of two game players then <img src="https://render.githubusercontent.com/render/math?math=g_j(A_i)"> is the gain of the row player when she selects the strategy *i* and the column player selects the strategy <img src="https://render.githubusercontent.com/render/math?math=C_j">. 
 
 
-The package mainly focuses on solving these kind of decision problems.
+The package mainly focuses on solving these kinds of decision problems.
 
 ## For whom?
 
-Multiple-criteria decision-making is an inter-discipline subject and there is a vast amount of research in the literature in this area. However, the existing software packages in this area are generally focused on a small subset of tools. JMcDM is a developer and researcher friendly Julia package that combines the developed methods, utility functions for implementing new ones, and serves an environment for comparing results of multiple analysis.  
+Multiple-criteria decision-making is an inter-discipline subject and there is a vast amount of research in the literature in this area. However, the existing software packages in this area are generally focused on a small subset of tools. JMcDM is a developer and researcher-friendly Julia package that combines the developed methods, utility functions for implementing new ones, and serves an environment for comparing results of multiple analyses.  
 
 ## Installation
 
@@ -48,7 +48,7 @@ in Julia REPL.
 
 ## Package Dependencies
 
-Since the Julia package manager installs all of the dependencies automatically, a standard user don't need to
+Since the Julia package manager installs all of the dependencies automatically, a standard user doesn't need to
 install them manually. The package dependencies are listed below:
 - Cbc
 - DataFrames
@@ -65,7 +65,7 @@ Please check out the reference manual [here](https://jbytecode.github.io/JMcDM/d
 ### MCDM Tools
 
 - TOPSIS (Technique for Order Preference by Similarity to Ideal Solutions)
-- ELECTRE (Elemination and Choice Translating Reality)
+- ELECTRE (Elimination and Choice Translating Reality)
 - DEMATEL (The Decision Making Trial and Evaluation Laboratory)
[…]
