### 2021-09-27 #3430 @drvinceknight
@jbytecode this is looking almost there now.

I have made some minor type edits in this PR: https://github.com/jbytecode/JMcDM/pull/29

I note that a few references do not have a DOI, for example: https://doi.org/10.1287/mnsc.31.6.647, would you be able to chase down the remaining ones please.
