+++ paper/paper.bib
@@ -16,7 +16,8 @@
   author    = {Greco, Salvatore and Figueira, Jose and Ehrgott, Matthias},
   volume    = {37},
   year      = {2016},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi       = {10.1007/978-1-4939-3094-4}
 }
 
 @article{PyTOPS,
@@ -135,7 +136,8 @@ and Vansnick, Jean-Claude},
   number    = {6},
   pages     = {647--656},
   year      = {1985},
-  publisher = {INFORMS}
+  publisher = {INFORMS},
+  doi       = {10.1287/mnsc.31.6.647}
 }
 
 
@@ -244,7 +246,8 @@ and Vansnick, Jean-Claude},
   number    = {2},
   pages     = {159--172},
   year      = {2010},
-  publisher = {Taylor \& Francis}
+  publisher = {Taylor \& Francis},
+  doi       = {10.3846/tede.2010.10}
 }
 
 
@@ -309,7 +312,8 @@ and Vansnick, Jean-Claude},
   author    = {Pamu{\v{c}}ar, Dragan and Vasin, Ljubislav and Lukovac, L},
   booktitle = {XVI international scientific-expert conference on railway, railcon},
   pages     = {89--92},
-  year      = {2014}
+  year      = {2014},
+  doi       = {10.13140/2.1.2707.6807}
 }
 
 @misc{copras,
