### 2021-09-28 #3430 @jbytecode
@drvinceknight

- your patch for the paper is merged in the main branch
- added some missing doi entries: [commit](https://github.com/jbytecode/JMcDM/commit/cc6327c64261488a519aa41850f38672c75ce9de)

Although this is not our first attempt, we can't still find doi entries for some other bibtex entries.
