### 2021-09-28 #3430 @jbytecode
After some modifications, it seems the layout has to be refactored because the table flows to the bottom of the page (maybe by adding a table reference).
