### 2021-09-28 #3430 @jbytecode
- Table is labeled and referenced in the text so layout seems to be okay.
