### 2021-09-28 #3430 @drvinceknight
Everything looks good to me. I've tried to find DOIs for the remaining 3 papers myself with no success.
