### 2021-09-29 #3430 @danielskatz
👋 @jbytecode - I've suggested some small changes in the paper in https://github.com/jbytecode/JMcDM/pull/30 - please merge this, or let me know what you disagree with, then we can proceed to final acceptance and publishing.
